Move room-related lookups into their own file.
There are some `room_*` functions that appear similar in nature
and functionality to their `hash_*` equivalents but which are not
actually implemented with hash tables (they use an array-based
lookup index instead); this moves those into their own file,
leaving only hash table-related items in `hash.[ch]`.
